Started work on fixing final refresh token bug.
