+++ src/main.ts
@@ -6,62 +6,12 @@ import { loadFonts } from "./plugins/webfontloader";
 import { createPinia } from "pinia";
 import { jwtDecode } from "jwt-decode";
 import './assets/main.css';
-import axios from "axios";
 import i18n, {defaultLocale, supportedLocales} from './i18n';
-
-import AuthenticationService from "./services/AuthenticationService";
 
 const pinia = createPinia();
 
 loadFonts();
 i18n.setup();
-
-// Configure axios to always include JWT and JWT-fingerprint when sending a request
-axios.defaults.withCredentials = true;
-axios.interceptors.request.use(
-    (config) => {
-      const jwt: string | null = sessionStorage.getItem("jwt");
-      if (jwt) {
-        if (config.headers) {
-          config.headers.Authorization = `Bearer ${jwt}`;
-        }
-      }
-      return config;
-    },
-    (error) => {
-      return Promise.reject(error);
-    }
-);
-
-axios.interceptors.response.use(
-  (response) => {
-    // Pass the response
-    return response;
-  },
-  async (error) => {
-    const originalRequest = error.config;
-    if (!originalRequest.url.endsWith("/user/authenticate") && !originalRequest.url.endsWith("/user/takeRole") && error.response) {
-      if (error.response.status === 401 && !originalRequest._retry) {
-        originalRequest._retry = true;
-  
-        const refreshToken: string | null = sessionStorage.getItem("refreshToken");
-        if (refreshToken) {
-          try {
-            const response = await AuthenticationService.refreshToken({refreshTokenValue: refreshToken});
-            sessionStorage.setItem("jwt", response.data.token);
-            sessionStorage.setItem("refreshToken", response.data.refreshToken);
-  
-            return axios(originalRequest);
-          } catch (refreshError) {
-            return Promise.reject(refreshError);
-          }
-        }
-      }
-    }
-
-    return Promise.reject(error);
-  }
-);
 
 router.beforeEach((to: any, from: any, next: any) => {
     const newLocale: string = to.params.locale;
