+++ src/stores/refreshTokenStore.ts
@@ -1,0 +1,15 @@
+import { defineStore } from 'pinia';
+
+export const useRefreshTokenStore = defineStore('refreshToken', {
+  state: () => ({
+    refreshing: false,
+  }),
+  actions: {
+    startedRefresh() {
+      this.refreshing = true;
+    },
+    finishedRefresh() {
+      this.refreshing = false;
+    },
+  },
+});
